Added unit tests

+++ src/main/java/lv/proofit/busapp/api/draft/price/request/Passenger.java
@@ -18,7 +18,7 @@ public class Passenger {
 
     @Schema(example = "2")
     @PositiveOrZero
-    BigDecimal numberOfBags;
+    int numberOfBags;
 
     @NotNull
     PassengerAge age;
